+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.5.2 (TBA)
+* Improvement: Renamed trading orderbook to orderbook
 * Bugfix: Fixed typo in funding capability (closes #1)
 
 # Version 0.5.1 (06.01.2018)